pyproject.toml
[tool.poetry]
name = "hack-large-text"
version = "0.1.0"
description = ""
authors = ["Yuriy Arabskyy <yuriy.arabskyy@gmail.com>"]

[[tool.poetry.packages]]
include = "hack"
from = "src"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
openai-agents = ">=0.3.3,<0.4.0"
pymupdf = "^1.26.4"
ipykernel = "^6.30.1"
sentence-transformers = "^5.1.1"
python-dotenv = "^1.1.1"
dspy-ai = "^3.0.3"
faiss-cpu = "^1.12.0"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.13.2"
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
